# src/main/resources/openapi/BikeHub.yaml
openapi: "3.1.0"
info:
  title: BikeHubParthenope
  version: "1.0"
servers:
  - url: http://localhost:8080
paths:
  /auth/signUp:
    post:
      summary: SignUp.
      description: Permette la registrazione.
      operationId: SignUp
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUp'
      responses:
        '201':
          description: Registrazione avvenuta con successo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/login:
    post:
      summary: User Login
      description: Permette ad un utente registrato di autenticarsi fornendo le credenziali (email e password) e ottenere un token di accesso.
      operationId: login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
      responses:
        '200':
          description: Autenticazione avvenuta con successo. Restituisce un token di accesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLogin'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /bicycle:
    post:
      description: Create bicycle.
      operationId: createBicycle
      tags:
        - Bicycle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBicycle'
      responses:
        '201':
          description: Bicycle successfully added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bicycle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /bicycle/{bicycleId}:
    get:
      description: Retrieve bicycle.
      operationId: retrieveBicycle
      tags:
        - Bicycle
      parameters:
        - in: path
          name: bicycleId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bicycle successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bicycle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      description: Update bicycle.
      operationId: updateBicycle
      tags:
        - Bicycle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBicycle'
      parameters:
        - in: path
          name: bicycleId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bicycle successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bicycle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      description: Delete bicycle.
      operationId: deleteBicycle
      tags:
        - Bicycle
      parameters:
        - in: path
          name: bicycleId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Bicycle successfully deleted.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /equipment:
    post:
      description: Create equipment.
      operationId: createEquipment
      tags:
        - Equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEquipment'
      responses:
        '201':
          description: Equipment successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /equipment/{equipmentId}:
    get:
      description: Retrieve equipment.
      operationId: retrieveEquipment
      tags:
        - Equipment
      parameters:
        - in: path
          name: equipmentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bicycle successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      description: Update equipment.
      operationId: updateEquipment
      tags:
        - Equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEquipment'
      parameters:
        - in: path
          name: equipmentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Equipment successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      description: Delete equipment.
      operationId: deleteEquipment
      tags:
        - Equipment
      parameters:
        - in: path
          name: equipmentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Equipment successfully deleted.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /rental:
    post:
      description: Create rental.
      operationId: createRental
      tags:
        - Rental
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRental'
      responses:
        '201':
          description: Rental successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rental'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /rental/{rentalId}:
    get:
      description: Retrieve rental.
      operationId: retrieveRental
      tags:
        - Rental
      parameters:
        - in: path
          name: rentalId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Rental successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rental'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      description: Update rental.
      operationId: updateRental
      tags:
        - Rental
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRental'
      parameters:
        - in: path
          name: rentalId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Rental successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rental'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      description: Delete rental.
      operationId: deleteRental
      tags:
        - Rental
      parameters:
        - in: path
          name: rentalId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Rental successfully deleted.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /rental/{rentalId}/pickup:
    put:
      description: Start a rental.
      operationId: pickupRental
      tags:
        - Rental
      parameters:
        - in: path
          name: rentalId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Rental successfully started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rental'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /rental/{rentalId}/return:
    put:
      description: Finish a rental.
      operationId: returnRental
      tags:
        - Rental
      requestBody:
        description: Dettagli per la riconsegna (parcheggio di restituzione)
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rental'
      parameters:
        - in: path
          name: rentalId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Rental successfully returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rental'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'


components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp (ISO 8601).
        status:
          type: integer
          format: int32
          description: HTTP status code (es. 400, 401, 500).
        error:
          type: string
          description: Error name (es. "BadRequest", "Unauthorized").
        message:
          type: string
          description: Detailed message.
        path:
          type: string
          description: Endpoint path that generated the error.
    SignUp:
      type: object
      required:
        - username
        - password
        - email
      properties:
        email:
          type: string
          format: email
          x-field-extra-annotation: "@NotBlank"
        password:
          type: string
        phoneNumber:
          type: string
        username:
          type: string
          maxLength: 32
          minLength: 3
        role:
          $ref: '#/components/schemas/AccountRole'
    Login:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    AuthLogin:
      type: object
      required:
        - accessToken
        - tokenType
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          default: Bearer
    CreateBicycle:
      type: object
      required:
        - currentParkingLotName
        - categories
        - chassisId
        - brand
        - model
        - hourlyPrice
      properties:
        currentParkingLotName:
          type: string
          x-field-extra-annotation: "@NotBlank"
        categories:
          type: array
          items:
            $ref: '#/components/schemas/BicycleCategory'
          uniqueItems: true
        chassisId:
          type: string
          x-field-extra-annotation: "@NotBlank"
        brand:
          type: string
          x-field-extra-annotation: "@NotBlank"
        model:
          type: string
          x-field-extra-annotation: "@NotBlank"
        hourlyPrice:
          type: number
          format: float
    Bicycle:
      type: object
      required:
        - currentParkingLotName
        - chassisId
        - status
        - brand
        - model
        - hourlyPrice
        - registeredAt
      properties:
        currentParkingLotName:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/BicycleCategory'
          uniqueItems: true
        equipments:
          type: array
          items:
            type: string
            format: uuid
        chassisId:
          type: string
        status:
          $ref: '#/components/schemas/BicycleStatus'
        brand:
          type: string
        model:
          type: string
        totalRentTime:
          type: number
          format: float
        hourlyPrice:
          type: number
          format: float
        registeredAt:
          type: string
          format: date-time
    UpdateBicycle:
      type: object
      properties:
        currentParkingLotName:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/BicycleCategory'
          uniqueItems: true
        status:
          $ref: '#/components/schemas/BicycleStatus'
        brand:
          type: string
        model:
          type: string
        hourlyPrice:
          type: number
          format: float
    CreateEquipment:
      type: object
      required:
        - type
        - name
        - description
      properties:
        bicycleId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/EquipmentType'
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
    Equipment:
      type: object
      required:
        - type
        - name
        - description
      properties:
        bicycleId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/EquipmentType'
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
    UpdateEquipment:
      type: object
      properties:
        bicycleId:
          type: string
          format: uuid
          nullable: true
        type:
          $ref: '#/components/schemas/EquipmentType'
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
          nullable: true
    CreateRental:
      type: object
      required:
        - bicycleId
        - requestedPickupTime
      properties:
        bicycleId:
          type: string
        requestedPickupTime:
          type: string
          format: date-time
    Rental:
      type: object
      required:
        - id
        - bicycleId
        - pickUpParkingLotName
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        bicycleId:
          type: string
        pickUpParkingLotName:
          type: string
        returnParkingLotName:
          type: string
        status:
          $ref: '#/components/schemas/RentalStatus'
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        totalPrice:
          type: number
          format: float
        mileage:
          type: number
          format: float
    UpdateRental:
      type: object
      properties:
        returnParkingLotName:
          type: string
        status:
          $ref: '#/components/schemas/RentalStatus'
        totalPrice:
          type: number
          format: float
        mileage:
          type: number
          format: float
    BicycleCategory:
      type: string
      enum:
        - ROAD
        - MOUNTAIN
        - ELECTRIC
        - BMX
        - RACING
    EquipmentType:
      type: string
      enum:
        - HELMET
        - GLOVES
        - GLASSES
    BicycleStatus:
      type: string
      enum:
        - AVAILABLE
        - RENTED
        - UNDER_MAINTENANCE
    AccountRole:
      type: string
      default: CUSTOMER
      enum:
        - CUSTOMER
        - ADMIN
    RentalStatus:
      type: string
      enum:
        - CREATED
        - IN_PROGRESS
        - FINISHED
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Autenticazione tramite Bearer token JWT nell'header 'Authorization'.